backend/requirements.txt
django==2.2.16
django-filter==2.3.0
djangorestframework==3.11.2
Pillow==8.3.1
webcolors==1.11.1
djoser==2.0.1